requirements.txt
requests
mysql-connector-python
streamlit